Changed our use of bigint id's to uuid's.
With the new requirement, NF2, that our client user interface be responsive even when the connection to the server isn't great, it would be much more useful to be able to create entity id's on the client rather than on the server.

We can look at the specific uuid version as part of implementation.