Created User schema and types.  More work needed.
Note that for the moment I'm just copying the same structure as all other types, however User is a special case as we most often don't want to include salt and password in our User type, we don't want to pass around their username, and so on.

But we'll flesh this out and refine it as we develop the application rather than second-guessing at this stage.